Updating the with-msw example to use TypeScript
