+++ examples/with-msw/.gitignore
@@ -7,6 +7,7 @@
 
 # testing
 /coverage
+public/mockServiceWorker.js
 
 # next.js
 /.next/
