+++ examples/with-msw/mocks/browser.ts
