+++ examples/with-msw/mocks/index.ts
@@ -9,3 +9,5 @@ async function initMocks() {
 }
 
 initMocks()
+
+export {};
