+++ examples/with-msw/mocks/types.ts
@@ -1,0 +1,11 @@
+export type Book = {
+  title: string
+  description: string
+  imageUrl: string
+}
+
+export type Review = {
+  id: string
+  author: string
+  text: string
+}
