+++ examples/with-msw/package.json
@@ -6,10 +6,16 @@
     "start": "next start"
   },
   "dependencies": {
-    "msw": "^0.29.0",
+    "msw": "^0.44.2",
     "next": "latest",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
+  "devDependencies": {
+    "@types/node": "18.6.3",
+    "@types/react": "18.0.15",
+    "@types/react-dom": "18.0.6",
+    "typescript": "^4.7.4"
   },
   "msw": {
     "workerDirectory": "public"
