+++ examples/with-msw/pages/_app.tsx
@@ -1,7 +1,9 @@
+import { AppProps } from "next/app"
+
 if (process.env.NEXT_PUBLIC_API_MOCKING === 'enabled') {
   require('../mocks')
 }
 
-export default function App({ Component, pageProps }) {
+export default function App({ Component, pageProps }: AppProps) {
   return <Component {...pageProps} />
 }
